Revert "Sys_monitoring page and agent approval component updated"
